+++ intercom_flutter_platform_interface/lib/method_channel_intercom_flutter.dart
@@ -236,6 +236,11 @@ class MethodChannelIntercomFlutter extends IntercomFlutterPlatform {
     await _channel.invokeMethod('displayTickets');
   }
 
+  @override
+  Future<void> displayHome() async {
+    await _channel.invokeMethod('displayHome');
+  }
+
   /// Convert the [PlatformException] details to [IntercomError].
   /// From the Platform side if the intercom operation failed then error details
   /// will be sent as details in [PlatformException].