+++ intercom_flutter_web/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.3
+
+* Implemented method `displayHome`.
+
 ## 1.1.2
 
 * Bump `web` to `^1.0.0`